Enrich Durability with each phase, so we can both prune unapplied dependencies to mitigate replicas that are behind causing Deps growth, and more reliably avoid initialising recovery progress log state of transactions that cannot yet make progress Also fix:  - Infer invalid in MaybeRecover and FetchData  - MaybeRecover sometimes aborts before home shard knows outcome  - Epoch sync with VisibilitySyncPoint  - Retired implies synced Also improve:  - Add Topology.removedNodes  - If Durability implies we can fetch a status, update the waiting state to fetch it
patch by Benedict; reviewed by Alex Petrov for CASSANDRA-20832

+++ accord-core/src/test/java/accord/local/RedundantBeforeTest.java
@@ -195,9 +195,9 @@ public class RedundantBeforeTest
         else return e.getValue().mergedLt;
     }
 
-    private static short merge(short a, short b)
+    private static int merge(int a, int b)
     {
-        return (short) (a | b);
+        return a | b;
     }
 
     private static SomeStatus merge(SomeStatus a, SomeStatus b)
